ファイル整備

+++ app/src/main/java/com/hamu/seihekijinrou/MeetingandVotingandResult/Online/Voting/onlineVoting3.kt
@@ -1,4 +1,4 @@
-package com.hamu.seihekijinrou
+package com.hamu.seihekijinrou.MeetingandVotingandResult.Online.Voting
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -12,21 +12,19 @@ import androidx.preference.PreferenceManager
 import com.google.firebase.firestore.SetOptions
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
-import com.hamu.seihekijinrou.MeetingandVotingandResult.Voting.abstractVoting
-import com.hamu.seihekijinrou.databinding.FragmentOnlineVoting2Binding
+import com.hamu.seihekijinrou.R
 import com.hamu.seihekijinrou.databinding.FragmentOnlineVoting3Binding
 
-
-class onlineVoting2 : OnlineabstractVoting() {
-    private var _binding: FragmentOnlineVoting2Binding?= null
+class onlineVoting3 : OnlineabstractVoting() {
+    private var _binding: FragmentOnlineVoting3Binding? = null
     private val binding get() = _binding!!
 
 
     override fun onCreateView(
             inflater: LayoutInflater, container: ViewGroup?,
-            savedInstanceState: Bundle?,
+            savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentOnlineVoting2Binding.inflate(inflater, container, false)
+        _binding = FragmentOnlineVoting3Binding.inflate(inflater, container, false)
         var pref = PreferenceManager.getDefaultSharedPreferences(context)
         var roomname = pref.getString("roomname", "")
         var numberofpeople = pref.getString("numberofpeople", "")
@@ -45,15 +43,20 @@ class onlineVoting2 : OnlineabstractVoting() {
             members = tmp.toMutableList()
             candidate1 = members[0]
             candidate2 = members[1]
+            candidate3 = members[2]
 
             binding.name1.text = candidate1
             binding.name2.text = candidate2
+            binding.name3.text = candidate3
         }
         binding.radioGroup.setOnCheckedChangeListener { group, checkedId ->
             when (checkedId) {
                 R.id.name1 -> Voted = candidate1
 
                 R.id.name2 -> Voted = candidate2
+
+                R.id.name3 -> Voted = candidate3
+
 
             }
 
@@ -148,27 +151,44 @@ class onlineVoting2 : OnlineabstractVoting() {
 
                 var vote1count = list1.count { it == candidate1 }
                 var vote2count = list1.count { it == candidate2 }
+                var vote3count = list1.count { it == candidate3 }
 
 
                 var list2 = mutableListOf<votedata>()
                 list2.add(votedata(candidate1, vote1count))
                 list2.add(votedata(candidate2, vote2count))
+                list2.add(votedata(candidate3, vote3count))
 
 
                 var list = list2.sortedByDescending { it.count }
 
 
 
-                if (list[0].count == list[1].count ) {
+                if (list[0].count == list[1].count && list[1].count == list[2].count) {
                     var remainmembers = members.toSet()
                     pref.edit {
                         putStringSet("remainmembers", remainmembers)
                     }.apply {}
                     Voting.delete()
                     whensameNumVoting()
+                } else if (list[0].count == list[1].count) {
+                    var remainmembers = setOf(list[2].name)
+                    var Suspectmembers = setOf(list[0].name, list[1].name)
+                    var pref = PreferenceManager.getDefaultSharedPreferences(context)
+                    pref.edit {
+                        putStringSet("remainmembers", remainmembers)
+                        putStringSet("Suspectmembers", Suspectmembers)
+                    }.apply {}
+                    if (Suspectmembers.contains(jinrouname)) {
+                        whendisagreeButContainjinrou()
+                    } else {
+                        whendisagree()
+                    }
+
+
                 } else {
                     Suspect = list[0].name
-                    var remainmembers = setOf(list[1].name)
+                    var remainmembers = setOf(list[1].name, list[2].name, list[3].name)
 
                     var pref = PreferenceManager.getDefaultSharedPreferences(context)
                     pref.edit {
@@ -184,15 +204,27 @@ class onlineVoting2 : OnlineabstractVoting() {
     }
 
     fun whensameNumVoting() {
-        var bundle = bundleOf("where" to 2)
+        var bundle = bundleOf("where" to 3)
 
-        findNavController().navigate(R.id.action_onlineVoting2_to_equalvote2, bundle)
+        findNavController().navigate(R.id.action_onlineVoting3_to_equalvote2, bundle)
     }
+
+    fun whendisagree() {
+
+        findNavController().navigate(R.id.action_onlineVoting3_to_whendisagree)
+    }
+
+    fun whendisagreeButContainjinrou() {
+
+        findNavController().navigate(R.id.action_onlineVoting3_to_whendisagreeButcontainJInrou)
+
+    }
+
     fun whenOpinionsAreUnited() {
 
 
         var bundle = bundleOf("Suspect" to Suspect)
 
-        findNavController().navigate(R.id.action_onlineVoting2_to_whenOpinionsAreUited, bundle)
+        findNavController().navigate(R.id.action_onlineVoting3_to_whenOpinionsAreUited, bundle)
     }
 }
